Conflicts:
	VERSION
	install.rdf

+++ content/controllers/ledger_controller.js
@@ -85,6 +85,13 @@
                                errmsg: _('An error was encountered while expiring backup ledger activity logs (error code %S) [messages #801].', [model.lastError])};
                     }
 
+                    r = model.execute('delete from order_payments where order_id IN (select id from ledger_records where created <= ' + expireDate + ')');
+                    if (!r) {
+                        throw {errno: model.lastError,
+                               errstr: model.lastErrorString,
+                               errmsg: _('An error was encountered while expiring ledger payment records (error code %S) [messages #803].', [model.lastError])};
+                    }
+
                     r = model.execute('delete from ledger_records where created <= ' + expireDate);
                     if (!r) {
                         throw {errno: model.lastError,
